enhetstester

+++ src/main/java/org/example/pensionat_backend/Controller/CustomerController.java
@@ -5,6 +5,7 @@ import org.example.pensionat_backend.DTO.CustomerDTO;
 import org.example.pensionat_backend.Models.Customer;
 import org.example.pensionat_backend.Repository.CustomerRepository;
 import org.example.pensionat_backend.Service.CustomerService;
+import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.validation.BindingResult;
@@ -37,12 +38,41 @@ public class CustomerController {
         if (result.hasErrors()) {
             return "CustomerReg"; // visa formuläret igen
         }
-        CustomerDTO savedDto = customerService.save(dto);
-        model.addAttribute("name", savedDto.getName());
-        model.addAttribute("email", savedDto.getEmail());
-        model.addAttribute("phone", savedDto.getPhone());
-        model.addAttribute("ssn", savedDto.getSsn());
-        model.addAttribute("id", savedDto.getId());
+
+        try {
+            CustomerDTO savedDto = customerService.save(dto);
+            model.addAttribute("name", savedDto.getName());
+            model.addAttribute("email", savedDto.getEmail());
+            model.addAttribute("phone", savedDto.getPhone());
+            model.addAttribute("ssn", savedDto.getSsn());
+            model.addAttribute("id", savedDto.getId());
+        } catch (DataIntegrityViolationException e) {
+            String message = e.getRootCause().getMessage();
+            boolean errorFound = false;
+
+            if (message.contains("uk_customer_name")) {
+                result.rejectValue("name", "error.user", "Namnet finns redan i databasen");
+                errorFound = true;
+            }
+            if (message.contains("uk_customer_ssn")) {
+                result.rejectValue("ssn", "error.user", "Det personnumret finns redan i databasen");
+                errorFound = true;
+            }
+            if (message.contains("uk_customer_email")) {
+                result.rejectValue("email", "error.user", "Email finns redan i databasen");
+                errorFound = true;
+            }
+            if (message.contains("uk_customer_phone")) {
+                result.rejectValue("phone", "error.user", "Telefonnumret finns redan i databasen");
+                errorFound = true;
+            }
+
+            if (!errorFound) {
+                result.reject("error.user", "oväntat fel uppstod");
+
+            }
+            return "CustomerReg";
+        }
 
         return "Welcome";
     }
@@ -63,19 +93,26 @@ public class CustomerController {
     }
 
     @GetMapping("/Customerlist")
-    public String listCustomers(Model model) {
+    public String listCustomers(Model model, @RequestParam(value = "error", required = false) String error) {
         List<CustomerDTO> customers = customerService.findAll();
         model.addAttribute("customers", customers);
+        if(error != null){
+            model.addAttribute("errorMessage", error);
+        }
         return "CustomerList";
     }
 
     @PostMapping("/deleteCustomer/{id}")
-    public String deleteCustomer(@PathVariable Long id) {
-        boolean deleted = customerService.deleteById(id);
-        if (!deleted) {
+    public String deleteCustomer(@PathVariable Long id, Model model) {
+        boolean customerHasBooking = customerService.customerHasBooking(id);
+        if (customerHasBooking) {
             return "redirect:/html/Customerlist?error=Kunden har aktiva bokningar och kan inte tas bort.";
-        }
+        } else {
+            customerService.deleteById(id);
         return "redirect:/html/Customerlist";
+
+        }
+
     }
 
     @PostMapping("/customer/edit/{id}")
@@ -83,13 +120,13 @@ public class CustomerController {
         Optional<CustomerDTO> customerDTO = Optional.of(new CustomerDTO());
         customerDTO = customerService.findById(id);
         if (customerDTO.isPresent()) {
-        model.addAttribute("customer", customerDTO.get());
-        return "editCustomer";
-        }
-        else {
+            model.addAttribute("customer", customerDTO.get());
+            return "editCustomer";
+        } else {
             return "redirect:/html/Customerlist?error=kund%20hittades%20inte";
         }
     }
+
     @PostMapping("/customer/edit")
     public String saveEditCustomer(@Valid @ModelAttribute("customer") CustomerDTO dto, BindingResult result, Model model) {
         if (result.hasErrors()) {
